+++ api/tests/shared/integration/infrastructure/jobs/campaign-result/ParticipationResultCalculationJobHandler_test.js
@@ -1,6 +1,12 @@
-import { CampaignParticipationStatuses, CampaignTypes } from '../../../../../lib/domain/models/index.js';
-import { ParticipationResultCalculationJobHandler } from '../../../../../src/shared/infrastructure/jobs/campaign-result/ParticipationResultCalculationJobHandler.js';
-import { databaseBuilder, expect, knex, learningContentBuilder, mockLearningContent } from '../../../../test-helper.js';
+import { CampaignParticipationStatuses, CampaignTypes } from '../../../../../../lib/domain/models/index.js';
+import { ParticipationResultCalculationJobHandler } from '../../../../../../src/shared/infrastructure/jobs/campaign-result/ParticipationResultCalculationJobHandler.js';
+import {
+  databaseBuilder,
+  expect,
+  knex,
+  learningContentBuilder,
+  mockLearningContent,
+} from '../../../../../test-helper.js';
 
 describe('Integration | Infrastructure | Jobs | CampaignResult | ParticipationResultCalculationJobHandler', function () {
   describe('#handle', function () {
