+++ api/tests/shared/integration/infrastructure/jobs/campaign-result/ParticipationResultCalculationJob_test.js
@@ -1,5 +1,5 @@
-import { ParticipationResultCalculationJob } from '../../../../../src/shared/infrastructure/jobs/campaign-result/ParticipationResultCalculationJob.js';
-import { expect, knex } from '../../../../test-helper.js';
+import { ParticipationResultCalculationJob } from '../../../../../../src/shared/infrastructure/jobs/campaign-result/ParticipationResultCalculationJob.js';
+import { expect, knex } from '../../../../../test-helper.js';
 
 describe('Integration | Infrastructure | Jobs | CampaignResult | ParticipationResultCalculation', function () {
   describe('#schedule', function () {
