+++ api/tests/shared/integration/infrastructure/jobs/campaign-result/SendSharedParticipationResultsToPoleEmploiJob_test.js
@@ -1,5 +1,5 @@
-import { SendSharedParticipationResultsToPoleEmploiJob } from '../../../../../src/shared/infrastructure/jobs/campaign-result/SendSharedParticipationResultsToPoleEmploiJob.js';
-import { expect, knex } from '../../../../test-helper.js';
+import { SendSharedParticipationResultsToPoleEmploiJob } from '../../../../../../src/shared/infrastructure/jobs/campaign-result/SendSharedParticipationResultsToPoleEmploiJob.js';
+import { expect, knex } from '../../../../../test-helper.js';
 
 describe('Integration | Infrastructure | Jobs | CampaignResult | SendSharedParticipationResultsToPoleEmploiJob', function () {
   describe('#schedule', function () {
