+++ api/tests/shared/unit/infrastructure/jobs/campaign-result/ParticipationResultCalculationJobHandler_test.js
@@ -1,7 +1,7 @@
-import { CampaignParticipationResultsShared } from '../../../../../lib/domain/events/CampaignParticipationResultsShared.js';
-import { usecases } from '../../../../../lib/domain/usecases/index.js';
-import { ParticipationResultCalculationJobHandler } from '../../../../../src/shared/infrastructure/jobs/campaign-result/ParticipationResultCalculationJobHandler.js';
-import { expect, sinon } from '../../../../test-helper.js';
+import { CampaignParticipationResultsShared } from '../../../../../../lib/domain/events/CampaignParticipationResultsShared.js';
+import { usecases } from '../../../../../../lib/domain/usecases/index.js';
+import { ParticipationResultCalculationJobHandler } from '../../../../../../src/shared/infrastructure/jobs/campaign-result/ParticipationResultCalculationJobHandler.js';
+import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | Infrastructure | Jobs | saveComputedCampaignParticipationResult', function () {
   context('#handle', function () {
