+++ api/tests/shared/unit/infrastructure/jobs/cpf-export/handlers/create-and-upload_test.js
@@ -6,8 +6,8 @@ import timezone from 'dayjs/plugin/timezone.js';
 import utc from 'dayjs/plugin/utc.js';
 import lodash from 'lodash';
 
-import { createAndUpload } from '../../../../../../src/shared/infrastructure/jobs/cpf-export/handlers/create-and-upload.js';
-import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
+import { createAndUpload } from '../../../../../../../src/shared/infrastructure/jobs/cpf-export/handlers/create-and-upload.js';
+import { domainBuilder, expect, sinon } from '../../../../../../test-helper.js';
 
 const { PassThrough, Readable } = stream;
