+++ api/tests/shared/unit/infrastructure/jobs/cpf-export/handlers/planner_test.js
@@ -1,8 +1,8 @@
 import dayjs from 'dayjs';
 
-import { config } from '../../../../../../src/shared/config.js';
-import { planner } from '../../../../../../src/shared/infrastructure/jobs/cpf-export/handlers/planner.js';
-import { expect, sinon } from '../../../../../test-helper.js';
+import { config } from '../../../../../../../src/shared/config.js';
+import { planner } from '../../../../../../../src/shared/infrastructure/jobs/cpf-export/handlers/planner.js';
+import { expect, sinon } from '../../../../../../test-helper.js';
 const { cpf } = config;
 import utc from 'dayjs/plugin/utc.js';
 import lodash from 'lodash';
