+++ api/tests/shared/unit/infrastructure/jobs/cpf-export/handlers/send-email_test.js
@@ -1,6 +1,6 @@
-import { sendEmail } from '../../../../../../src/shared/infrastructure/jobs/cpf-export/handlers/send-email.js';
-import { logger } from '../../../../../../src/shared/infrastructure/utils/logger.js';
-import { expect, sinon } from '../../../../../test-helper.js';
+import { sendEmail } from '../../../../../../../src/shared/infrastructure/jobs/cpf-export/handlers/send-email.js';
+import { logger } from '../../../../../../../src/shared/infrastructure/utils/logger.js';
+import { expect, sinon } from '../../../../../../test-helper.js';
 
 describe('Unit | Infrastructure | jobs | cpf-export | send-email', function () {
   let getPreSignedUrls;
