+++ api/tests/shared/unit/infrastructure/jobs/organization-learner/ScheduleComputeOrganizationLearnersCertificabilityJobHandler_test.js
@@ -2,8 +2,8 @@ import dayjs from 'dayjs';
 import timezone from 'dayjs/plugin/timezone.js';
 import utc from 'dayjs/plugin/utc.js';
 
-import { ScheduleComputeOrganizationLearnersCertificabilityJobHandler } from '../../../../../src/shared/infrastructure/jobs/organization-learner/ScheduleComputeOrganizationLearnersCertificabilityJobHandler.js';
-import { expect, knex, sinon } from '../../../../test-helper.js';
+import { ScheduleComputeOrganizationLearnersCertificabilityJobHandler } from '../../../../../../src/shared/infrastructure/jobs/organization-learner/ScheduleComputeOrganizationLearnersCertificabilityJobHandler.js';
+import { expect, knex, sinon } from '../../../../../test-helper.js';
 dayjs.extend(utc);
 dayjs.extend(timezone);
